feat: integrate none-ls-extras.nvim for enhanced formatting options
This commit introduces the `none-ls-extras.nvim` plugin to the Neovim configuration. The plugin is added as a dependency in the `plugins.lua` file and locked in the `lazy-lock.json` file.

Additionally, the `null-ls.lua` configuration is updated to utilize the `none-ls-extras.nvim` plugin for formatting with `golangci_lint`. Unused comments and redundant code related to formatting and actions are also cleaned up for better readability and maintainability.